+++ src/num/niche/non_value.rs
@@ -16,6 +16,7 @@ use crate::mem::{bit_sized, ByteSized};
 use crate::{
     _dep::_core::{fmt, num::*, str::FromStr},
     code::{paste, ConstDefault},
+    error::unwrap,
 };
 
 macro_rules! impl_non_value {
@@ -27,19 +28,28 @@ macro_rules! impl_non_value {
             8:"_non_value_u8", 16:"_non_value_u16", 32:"_non_value_u32",
             64:"_non_value_u64", 128:"_non_value_u128", size:"_non_value_usize"];
     };
-    ($abs:ident, $doc:literal, $s:ident, $( $b:tt: $cap:literal ),+) => {
+    ($XTR:ident, $doc:literal, $s:ident, $( $b:tt: $cap:literal ),+) => {
         $( paste!{
-            impl_non_value![@[<NonValue $s:upper $b>], $abs, $doc, $s, $b : $cap];
+            impl_non_value![@
+                [<NonValue $s:upper $b>], // $name
+                $XTR,
+                $doc,
+                [<$s $b>], // $IP
+                $s,
+                $b:
+                $cap
+            ];
         })+
     };
 
     // $name: the full name of the new type. E.g. NonValueI8.
-    // $abs:  the absolute maximum value constant for this type
+    // $XTR:  the *extreme* value constant for this type. (MIN | MAX).
     // $doc:  the specific beginning of the documentation.
+    // $IP:   the type of the corresponding integer primitive. E.g. i8
     // $s:    the sign identifier: i or u.
     // $b:    the bits of the type, from 8 to 128, or the `size` suffix.
     // $cap:  the capability feature that enables the given implementation. E.g "_non_value_i8".
-    (@$name:ident, $abs:ident, $doc:literal, $s:ident, $b:tt : $cap:literal) => { paste! {
+    (@$name:ident, $XTR:ident, $doc:literal, $IP:ty, $s:ident, $b:tt : $cap:literal) => { paste! {
         /* definition */
 
         #[doc = $doc " integer that is known not to equal some specific value." ]
@@ -59,18 +69,18 @@ macro_rules! impl_non_value {
         #[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
         #[cfg(feature = $cap )]
         #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = $cap)))]
-        pub struct $name <const V: [<$s $b>]>([<NonZero $s:upper $b>]);
+        pub struct $name <const V: $IP>([<NonZero $s:upper $b>]);
 
         /* aliases */
 
         #[doc = $doc " integer that is known not to equal its most extreme value ([`"
-            $abs "`][" [<$s $b>] "::" $abs "])."]
+            $XTR "`][" $IP "::" $XTR "])."]
         ///
         /// Unlike the `NonValue*` types in general, this type alias implements
         /// the [`Default`] and [`ConstDefault`][crate::code::ConstDefault] traits.
         #[cfg(feature = $cap )]
         #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = $cap)))]
-        pub type [<NonExtreme $s:upper $b>] = $name <{[<$s $b>]::$abs}>;
+        pub type [<NonExtreme $s:upper $b>] = $name <{$IP::$XTR}>;
 
         #[cfg(feature = $cap )]
         #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = $cap)))]
@@ -80,11 +90,11 @@ macro_rules! impl_non_value {
             #[inline] #[must_use]
             fn default() -> Self {
                 #[cfg(any(feature = "safe_num", not(feature = "unsafe_niche")))]
-                return [<NonExtreme $s:upper $b>]::new([<$s $b>]::default()).unwrap();
+                return [<NonExtreme $s:upper $b>]::new($IP::default()).unwrap();
 
                 #[cfg(all(not(feature = "safe_num"), feature = "unsafe_niche"))]
                 // SAFETY: the default primitive value is always 0, and their MAX is never 0.
-                unsafe { return [<NonExtreme $s:upper $b>]::new_unchecked([<$s $b>]::default()); }
+                unsafe { return [<NonExtreme $s:upper $b>]::new_unchecked($IP::default()); }
             }
         }
 
@@ -94,23 +104,44 @@ macro_rules! impl_non_value {
             /// Makes use of the `unsafe_niche` feature if enabled.
             const DEFAULT: Self = {
                 #[cfg(any(feature = "safe_num", not(feature = "unsafe_niche")))]
-                if let Some(v) = Self::new([<$s $b>]::DEFAULT) { v } else { unreachable![] }
+                if let Some(v) = Self::new($IP::DEFAULT) { v } else { unreachable![] }
 
                 #[cfg(all(not(feature = "safe_num"), feature = "unsafe_niche"))]
                 // SAFETY: the default primitive value is always 0, and their MAX is never 0.
-                unsafe { [<NonExtreme $s:upper $b>]::new_unchecked([<$s $b>]::DEFAULT) }
+                unsafe { [<NonExtreme $s:upper $b>]::new_unchecked($IP::DEFAULT) }
             };
         }
 
-        /* methods */
-
         #[cfg(feature = $cap )]
         #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = $cap)))]
-        impl<const V: [<$s $b>]> $name<V> {
+        impl<const V: $IP> $name<V> {
+            /* constants */
+
+            /// Returns the maximum possible value.
+            #[inline] #[must_use]
+            pub const MAX: Self = {
+                if $IP::MAX > V {
+                    unwrap![some Self::new($IP::MAX)]
+                } else {
+                    unwrap![some Self::new($IP::MAX - 1)]
+                }
+            };
+            /// Returns the minimum possible value.
+            #[inline] #[must_use]
+            pub const MIN: Self = {
+                if $IP::MIN < V {
+                    unwrap![some Self::new($IP::MIN)]
+                } else {
+                    unwrap![some Self::new($IP::MIN + 1)]
+                }
+            };
+
+            /* methods */
+
             #[doc = "Returns a `" $name "` with the given `value`,"
                 " if it is not equal to `V`."]
             #[must_use]
-            pub const fn new(value: [<$s $b>]) -> Option<Self> {
+            pub const fn new(value: $IP) -> Option<Self> {
                 match [<NonZero $s:upper $b>]::new(value ^ V) {
                     None => None,
                     Some(v) => Some(Self(v)),
@@ -126,7 +157,7 @@ macro_rules! impl_non_value {
             #[must_use]
             #[cfg(all(not(feature = "safe_num"), feature = "unsafe_niche"))]
             #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "unsafe_niche")))]
-            pub const unsafe fn new_unchecked(value: [<$s $b>]) -> Self {
+            pub const unsafe fn new_unchecked(value: $IP) -> Self {
                 #[cfg(debug_assertions)]
                 if value == V { panic!("The given value was specifically prohibited.") }
 
@@ -137,7 +168,7 @@ macro_rules! impl_non_value {
             /// Returns the value as a primitive type.
             #[inline]
             #[must_use]
-            pub const fn get(&self) -> [<$s $b>] {
+            pub const fn get(&self) -> $IP {
                 self.0.get() ^ V
             }
 
@@ -156,68 +187,68 @@ macro_rules! impl_non_value {
 
             /* core impls */
 
-            impl<const V: [<$s $b>]> fmt::Display for $name <V> {
+            impl<const V: $IP> fmt::Display for $name <V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     write!(f, "{}", self.get())
                 }
             }
-            impl<const V: [<$s $b>]> fmt::Debug for $name <V> {
+            impl<const V: $IP> fmt::Debug for $name <V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     write!(f, "{}::<{}>({})", stringify!($name), V, self.get())
                 }
             }
-            impl<const V: [<$s $b>]> fmt::Binary for $name<V> {
+            impl<const V: $IP> fmt::Binary for $name<V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     fmt::Binary::fmt(&self.get(), f)
                 }
             }
-            impl<const V: [<$s $b>]> fmt::Octal for $name<V> {
+            impl<const V: $IP> fmt::Octal for $name<V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     fmt::Octal::fmt(&self.get(), f)
                 }
             }
-            impl<const V: [<$s $b>]> fmt::LowerHex for $name<V> {
+            impl<const V: $IP> fmt::LowerHex for $name<V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     fmt::LowerHex::fmt(&self.get(), f)
                 }
             }
-            impl<const V: [<$s $b>]> fmt::UpperHex for $name<V> {
+            impl<const V: $IP> fmt::UpperHex for $name<V> {
                 #[inline]
                 fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                     fmt::UpperHex::fmt(&self.get(), f)
                 }
             }
 
-            impl<const V: [<$s $b>]> FromStr for $name<V> {
+            impl<const V: $IP> FromStr for $name<V> {
                 type Err = ParseIntError;
                 #[inline]
                 fn from_str(s: &str) -> Result<Self, Self::Err> {
-                    Self::new([<$s $b>]::from_str(s)?).ok_or_else(||"".parse::<i32>().unwrap_err())
+                    Self::new($IP::from_str(s)?).ok_or_else(||"".parse::<i32>().unwrap_err())
                 }
             }
 
             /* conversions */
 
-            impl<const V: [<$s $b>]> From<$name<V>> for [<$s $b>] {
+            impl<const V: $IP> From<$name<V>> for $IP {
                 #[inline]
                 #[must_use]
-                fn from(value: $name<V>) -> [<$s $b>] {
+                fn from(value: $name<V>) -> $IP {
                     value.get()
                 }
             }
 
-            impl<const V: [<$s $b>]> TryFrom<[<$s $b>]> for $name<V> {
+            impl<const V: $IP> TryFrom<$IP> for $name<V> {
                 type Error = core::num::TryFromIntError;
 
                 /// # Features
                 /// Makes use of the `unsafe_niche` feature if enabled.
                 #[inline]
-                fn try_from(value: [<$s $b>]) -> Result<Self, Self::Error> {
+                fn try_from(value: $IP) -> Result<Self, Self::Error> {
                     // We generate a TryFromIntError by intentionally causing a failed conversion.
                     #[cfg(any(feature = "safe_num", not(feature = "unsafe_niche")))]
                     return Self::new(value).ok_or_else(|| i8::try_from(255_u8).unwrap_err());
@@ -232,11 +263,11 @@ macro_rules! impl_non_value {
 
             // BitSized
             #[cfg(feature = "mem_bit")]
-            bit_sized![<const V: [<$s $b>]> =
-                { [<$s $b>]::BYTE_SIZE * 8}; for $name<V>];
+            bit_sized![<const V: $IP> =
+                { $IP::BYTE_SIZE * 8}; for $name<V>];
 
             #[cfg(feature = "unsafe_layout")]
-            unsafe impl<const V: [<$s $b>]> MemPod for Option<$name<V>> {}
+            unsafe impl<const V: $IP> MemPod for Option<$name<V>> {}
 
             /* external impls*/
 
@@ -247,11 +278,11 @@ macro_rules! impl_non_value {
             mod [<$name $s $b>] {
                 use super::*;
 
-                unsafe impl<const V: [<$s $b>]> ZeroableInOption for $name<V> {}
-                unsafe impl<const V: [<$s $b>]> PodInOption for $name<V> {}
-                unsafe impl<const V: [<$s $b>]> NoUninit for $name<V> {}
-                unsafe impl<const V: [<$s $b>]> CheckedBitPattern for $name<V> {
-                    type Bits = [<$s $b>];
+                unsafe impl<const V: $IP> ZeroableInOption for $name<V> {}
+                unsafe impl<const V: $IP> PodInOption for $name<V> {}
+                unsafe impl<const V: $IP> NoUninit for $name<V> {}
+                unsafe impl<const V: $IP> CheckedBitPattern for $name<V> {
+                    type Bits = $IP;
 
                     #[inline]
                     fn is_valid_bit_pattern(bits: &Self::Bits) -> bool {